+++ ui/mvc/Hubs/CosmosDocuments.cs
@@ -1,32 +1,18 @@
-﻿using Azure.Core;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.SignalR;
-using Microsoft.Azure.Cosmos;
-using PhiDeidPortal.Ui.Controllers;
+﻿using Microsoft.AspNetCore.SignalR;
+using PhiDeidPortal.Ui.Entities;
+using PhiDeidPortal.Ui.Interfaces;
 using PhiDeidPortal.Ui.Services;
-using PhiDeidPortal.Ui.ViewComponents;
-using System.Security.Claims;
-using IAuthorizationService = PhiDeidPortal.Ui.Services.IAuthorizationService;
 
 namespace PhiDeidPortal.Ui.Hubs
-{    
+{
     public class CosmosDocuments : Hub
     {
         private readonly ICosmosService _cosmosService;
-        private readonly IAuthorizationService _authService;
-        private readonly IAISearchService _indexQueryer;
-        private readonly CosmosClient _cosmosClient;
-        private readonly IFeatureService _featureService;
         private readonly IUserContextService _userContextService;
 
-        public CosmosDocuments(ICosmosService cosmosService, IAuthorizationService authService, IAISearchService indexQueryer, CosmosClient cosmosClient, IFeatureService featureService, IUserContextService userContextService)
+        public CosmosDocuments(ICosmosService cosmosService, IUserContextService userContextService)
         {
-            _authService = authService;
             _cosmosService = cosmosService;
-            _indexQueryer = indexQueryer;
-            _cosmosClient = cosmosClient;
-            _featureService = featureService;
             _userContextService = userContextService;
         }
 
@@ -46,32 +32,5 @@ namespace PhiDeidPortal.Ui.Hubs
             var summary = (isElevated && !viewFilter) ? _cosmosService.GetSummary() : _cosmosService.GetSummaryByAuthor(currentUser.Identity.Name);
             await Clients.All.SendAsync("UpdateCounts", summary);
         }
-    }
-
-    public interface IUserContextService
-    {
-        ClaimsPrincipal User { get; set; }
-        bool HasElevatedRights { get; set; }
-        bool ViewFilter { get; set; }
-    }
-
-    public class UserContextService : IUserContextService
-    {
-        public ClaimsPrincipal User { get; set; }
-        public bool HasElevatedRights { get; set; }
-        public bool ViewFilter { get; set; }
-    }
-
-    public class CosmosRecord
-    {
-        public string id { get; set; }
-        public string Uri { get; set; }
-        public string FileName { get; set; }
-        public string Status { get; set; }
-        public string Author { get; set; }
-        public bool AwaitingIndex { get; set; }
-        public string JustificationText { get; set; }
-        public DateTime LastIndexed { get; set; }
-        public string[] OrganizationalMetadata { get; set; }
     }
 }
