+++ ui/mvc/Interfaces/IUserContextService.cs
@@ -1,0 +1,11 @@
+﻿namespace PhiDeidPortal.Ui.Interfaces
+{
+    using System.Security.Claims;
+
+    public interface IUserContextService
+    {
+        ClaimsPrincipal User { get; set; }
+        bool HasElevatedRights { get; set; }
+        bool ViewFilter { get; set; }
+    }
+}
