+++ ui/mvc/Pages/Justification.cshtml
@@ -61,8 +61,7 @@
                                     {
                                         <button class="btn btn-danger delete-button" type="button" data-href='@result.Document["metadata_storage_path"]'><i class="bi bi-x-lg"></i> Delete</button>
                                     }
-                                }
-                                <span>&nbsp;Awaiting Indexing: @awaitingIndexing</span>
+                                }                                
                             </div>
                         </div>
                     </td>
