+++ ui/mvc/Services/UserContextService.cs
@@ -1,0 +1,12 @@
+﻿using PhiDeidPortal.Ui.Interfaces;
+using System.Security.Claims;
+
+namespace PhiDeidPortal.Ui.Services
+{
+    public class UserContextService : IUserContextService
+    {
+        public ClaimsPrincipal User { get; set; }
+        public bool HasElevatedRights { get; set; }
+        public bool ViewFilter { get; set; }
+    }
+}
